server: now marking food as ready will reduce the ingredients amount

+++ server/src/main/java/com/ifmo/ratatoile/controller/IngredientController.kt
@@ -32,6 +32,6 @@ class IngredientController(
   @GetMapping("/warehouse/{id}/{amount}")
   fun setIngredientAmount(
     @PathVariable("id") id: Int,
-    @PathVariable("amount") amount: Double
+    @PathVariable("amount") amount: Float
   ) = ingredientService.setAmount(id, amount)
 }
